requirements.txt
fastapi
uvicorn
pandas
scikit-learn
sqlalchemy
psycopg2-binary
python-dotenv